pokemon_app/requirements.txt
Flask==3.0.0
mysql-connector-python==8.1.0
Requests==2.31.0
